+++ src/ViennaRNA.jl
@@ -10,9 +10,9 @@ export FoldCompound, Pairtable
 # wrapper functions
 export bp_distance, bpp, centroid, energy, ensemble_defect,
     heat_capacity, inverse_fold, inverse_pf_fold, mea,
-    mean_bp_distance, mfe, mfe_window, neighbors, partfn, plot_coords,
-    prob_of_structure, sample_structures, subopt, subopt_zuker,
-    tree_edit_dist
+    mean_bp_distance, mfe, mfe_window, mfe_window_channel, neighbors,
+    partfn, plot_coords, prob_of_structure, sample_structures, subopt,
+    subopt_zuker, tree_edit_dist
 
 # additional utility functions
 export basepairs, prob_of_basepairs